+++ internal/elasticsearch/index/index/models.go
@@ -26,8 +26,8 @@ import (
 
 	estypes "github.com/elastic/go-elasticsearch/v8/typedapi/types"
 	"github.com/elastic/terraform-provider-elasticstack/internal/clients"
-	"github.com/elastic/terraform-provider-elasticstack/internal/clients/elasticsearch"
 	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/index"
+	"github.com/elastic/terraform-provider-elasticstack/internal/elasticsearch/index/aliasutil"
 	"github.com/elastic/terraform-provider-elasticstack/internal/models"
 	"github.com/elastic/terraform-provider-elasticstack/internal/utils/customtypes"
 	"github.com/elastic/terraform-provider-elasticstack/internal/utils/typeutils"
@@ -35,7 +35,6 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/attr"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
 	"github.com/hashicorp/terraform-plugin-framework/types"
-	"github.com/hashicorp/terraform-plugin-framework/types/basetypes"
 )
 
 var (
@@ -189,16 +188,6 @@ type tfModel struct {
 	Settings                           types.List           `tfsdk:"settings"`
 }
 
-type aliasTfModel struct {
-	Name          types.String         `tfsdk:"name"`
-	Filter        jsontypes.Normalized `tfsdk:"filter"`
-	IndexRouting  types.String         `tfsdk:"index_routing"`
-	IsHidden      types.Bool           `tfsdk:"is_hidden"`
-	IsWriteIndex  types.Bool           `tfsdk:"is_write_index"`
-	Routing       types.String         `tfsdk:"routing"`
-	SearchRouting types.String         `tfsdk:"search_routing"`
-}
-
 type settingsTfSet struct {
 	Setting types.Set `tfsdk:"setting"`
 }
@@ -226,7 +215,7 @@ func (model *tfModel) populateFromAPI(ctx context.Context, indexName string, api
 		model.Name = types.StringValue(indexName)
 	}
 
-	modelAliases, diags := aliasesFromAPI(ctx, apiModel)
+	modelAliases, diags := aliasutil.AliasesFromAPI(ctx, apiModel.Aliases, aliasElementType(ctx))
 	if diags.HasError() {
 		return diags
 	}
@@ -270,25 +259,6 @@ func (model *tfModel) populateFromAPI(ctx context.Context, indexName string, api
 	return nil
 }
 
-func aliasesFromAPI(ctx context.Context, apiModel estypes.IndexState) (basetypes.SetValue, diag.Diagnostics) {
-	aliases := []aliasTfModel{}
-	for name, alias := range apiModel.Aliases {
-		tfAlias, diags := newAliasModelFromAPI(name, alias)
-		if diags.HasError() {
-			return basetypes.SetValue{}, diags
-		}
-
-		aliases = append(aliases, tfAlias)
-	}
-
-	modelAliases, diags := types.SetValueFrom(ctx, aliasElementType(ctx), aliases)
-	if diags.HasError() {
-		return basetypes.SetValue{}, diags
-	}
-
-	return modelAliases, nil
-}
-
 func setSettingsFromAPI(model *tfModel, apiModel estypes.IndexState) diag.Diagnostics {
 	settingsBytes, err := json.Marshal(apiModel.Settings)
 	if err != nil {
@@ -316,14 +286,14 @@ func (model tfModel) toAPIModel(ctx context.Context) (models.Index, diag.Diagnos
 	if typeutils.IsKnown(model.Alias) {
 		apiModel.Aliases = map[string]models.IndexAlias{}
 
-		var planAliases []aliasTfModel
+		var planAliases []aliasutil.AliasModel
 		diags.Append(model.Alias.ElementsAs(ctx, &planAliases, true)...)
 		if diags.HasError() {
 			return models.Index{}, diags
 		}
 
 		for _, planAlias := range planAliases {
-			apiAlias, diags := planAlias.toAPIModel()
+			apiAlias, diags := aliasToAPIModel(planAlias)
 			if diags.HasError() {
 				return models.Index{}, diags
 			}
@@ -585,7 +555,7 @@ func convertSettingsKeyToTFFieldKey(settingKey string) string {
 	return strings.ReplaceAll(settingKey, ".", "_")
 }
 
-func (model aliasTfModel) toAPIModel() (models.IndexAlias, diag.Diagnostics) {
+func aliasToAPIModel(model aliasutil.AliasModel) (models.IndexAlias, diag.Diagnostics) {
 	apiModel := models.IndexAlias{
 		Name:          model.Name.ValueString(),
 		IndexRouting:  model.IndexRouting.ValueString(),
@@ -604,40 +574,6 @@ func (model aliasTfModel) toAPIModel() (models.IndexAlias, diag.Diagnostics) {
 	return apiModel, nil
 }
 
-func newAliasModelFromAPI(name string, apiModel estypes.Alias) (aliasTfModel, diag.Diagnostics) {
-	tfAlias := aliasTfModel{
-		Name:          types.StringValue(name),
-		IndexRouting:  types.StringValue(typeutils.Deref(apiModel.IndexRouting)),
-		IsHidden:      types.BoolValue(typeutils.Deref(apiModel.IsHidden)),
-		IsWriteIndex:  types.BoolValue(typeutils.Deref(apiModel.IsWriteIndex)),
-		Routing:       types.StringValue(typeutils.Deref(apiModel.Routing)),
-		SearchRouting: types.StringValue(typeutils.Deref(apiModel.SearchRouting)),
-	}
-
-	if apiModel.Filter != nil {
-		filterBytes, err := json.Marshal(apiModel.Filter)
-		if err != nil {
-			return aliasTfModel{}, diag.Diagnostics{
-				diag.NewErrorDiagnostic("failed to marshal alias filter", err.Error()),
-			}
-		}
-		var filterMap map[string]any
-		if err := json.Unmarshal(filterBytes, &filterMap); err != nil {
-			return aliasTfModel{}, diag.Diagnostics{
-				diag.NewErrorDiagnostic("failed to unmarshal alias filter", err.Error()),
-			}
-		}
-		normalized := elasticsearch.NormalizeQueryFilter(filterMap)
-		if nm, ok := normalized.(map[string]any); ok {
-			filterMap = nm
-		}
-		normalizedBytes, _ := json.Marshal(filterMap)
-		tfAlias.Filter = jsontypes.NewNormalizedValue(string(normalizedBytes))
-	}
-
-	return tfAlias, nil
-}
-
 func indexStateToModel(state estypes.IndexState) (models.Index, diag.Diagnostics) {
 	var model models.Index
 
@@ -653,19 +589,11 @@ func indexStateToModel(state estypes.IndexState) (models.Index, diag.Diagnostics
 				SearchRouting: typeutils.Deref(alias.SearchRouting),
 			}
 			if alias.Filter != nil {
-				filterBytes, err := json.Marshal(alias.Filter)
-				if err != nil {
-					return models.Index{}, diag.Diagnostics{
-						diag.NewErrorDiagnostic("failed to marshal alias filter", err.Error()),
-					}
+				filterMap, diags := aliasutil.NormalizeAliasFilterAnyToMap(alias.Filter)
+				if diags.HasError() {
+					return models.Index{}, diags
 				}
-				var filterMap map[string]any
-				if err := json.Unmarshal(filterBytes, &filterMap); err != nil {
-					return models.Index{}, diag.Diagnostics{
-						diag.NewErrorDiagnostic("failed to unmarshal alias filter", err.Error()),
-					}
-				}
-				indexAlias.Filter = elasticsearch.NormalizeQueryFilter(filterMap).(map[string]any)
+				indexAlias.Filter = filterMap
 			}
 			model.Aliases[name] = indexAlias
 		}